Fix importing of DX Gateway associations via DX Gateway import.
