+++ aws/resource_aws_dx_gateway.go
@@ -125,6 +125,28 @@ func resourceAwsDxGatewayDelete(d *schema.ResourceData, meta interface{}) error 
 	return nil
 }
 
+func resourceAwsDxGatewayImportState(d *schema.ResourceData, meta interface{}) ([]*schema.ResourceData, error) {
+	conn := meta.(*AWSClient).dxconn
+
+	resp, err := conn.DescribeDirectConnectGatewayAssociations(&directconnect.DescribeDirectConnectGatewayAssociationsInput{
+		DirectConnectGatewayId: aws.String(d.Id()),
+	})
+	if err != nil {
+		return nil, fmt.Errorf("error reading Direct Connect gateway association: %s", err)
+	}
+
+	results := []*schema.ResourceData{d}
+	for _, assoc := range resp.DirectConnectGatewayAssociations {
+		d := resourceAwsDxGatewayAssociation().Data(nil)
+		d.SetType("aws_dx_gateway_association")
+		d.SetId(dxGatewayAssociationId(aws.StringValue(assoc.DirectConnectGatewayId), aws.StringValue(assoc.AssociatedGateway.Id)))
+		d.Set("dx_gateway_association_id", assoc.AssociationId)
+		results = append(results, d)
+	}
+
+	return results, nil
+}
+
 func dxGatewayStateRefresh(conn *directconnect.DirectConnect, dxgwId string) resource.StateRefreshFunc {
 	return func() (interface{}, string, error) {
 		resp, err := conn.DescribeDirectConnectGateways(&directconnect.DescribeDirectConnectGatewaysInput{
